Clicking buttons
Found buttons to click the drop down for the calendar and to go to the previous month. Able to fetch the text from the divs to get the current month and year. Next need to go back to start of 2023-2024 season to begin parsing from there and look for divs with data.