Figured out how to insert dates
I think I've done everything I need to do with the database for now. Next step is to actually go through and parse a day's worth of games into the database. Also need to put database connection code in a different file which we do not commit to github and change the password to the database.