Created method to parse moneyline page
Created the method that takes in two parameters, the moneyline webpage element and the enum for the bookmaker we want to grab odds from. We go through an find all games and odds and return a list of moneyline objects. Next step is to write the method that inserts these into the database. Also made the database connection secure using an enum class which we will not commit. Still need to change password though.