Actually parsed data into database!
Parsed all moneyline data from the games that were played on 2023-11-07 with the odds from FANDUEL. We need to address the fact that all odds values are being inserted as positive integers. We need to get better control over messing with the calendar component of the website, preferably something that abstracts away the complexity to the point that we can just enter a date we want to pull games from and have it be done.