Almost finished with date finder
Almost finished with date finder method, just need to finish up debugging. After this we need to parse some more days in and see what bugs we run into. Also need to make decisions about partial insertions and whatnot and if we want to handle some exceptions and just not enter those games.