Started API
Started to create the API. Thought a lot about the design and how we will add a front end. API is very bare bones but next step is probably to finish out the scraping portion of the project so we can start running batch jobs that are scheduled to parse every hour or so.