Wrote parsing for spreads
Wrote the parsing algorithms for spreads, need to write the method to insert into the database still though. Next step is to do the same with over/unders, and then go to a different page to parse the results of the games. Once we have all this, we have collected all the data we want, and it is now time to work on the API and maybe think of better or more intuitive ways to store the data (foreign keys, etc.)