Found some bugs
We found a bug related to the createBlankFromTwoRows columns. When we grab the data from these rows, we need to use the "game-odds" classes. If we don't use these classes, and we try to use "data-odds" and "data-value", then on games where certain sportsbooks don't give odds our indexing will be all messed up. Next step is to fix these, and then work on inserts for spread and total.