Finished insert for spread and total
Finished the insert methods for spread and total. We can now sort of 'do' everything, we just need to fit it together and find a way to be able to parse like 1 day's worth of data every hour and create some logs for us. We really want to get it right the first time because it's gonna be annoying if we have to go back and delete a bunch of data because we messed something up in some of the initial fetchs.